+++ front/lib/actions/mcp_internal_actions/servers/notion/index.ts
@@ -11,10 +11,32 @@ import type {
 } from "@notionhq/client/build/src/api-endpoints";
 import { APIResponseError } from "@notionhq/client/build/src/errors";
 import { parseISO } from "date-fns";
-import { z } from "zod";
 
 import { MCPError } from "@app/lib/actions/mcp_errors";
 import type { SearchResultResourceType } from "@app/lib/actions/mcp_internal_actions/output_schemas";
+import {
+  addPageContentSchema,
+  createCommentSchema,
+  createDatabaseSchema,
+  createPageSchema,
+  deleteBlockSchema,
+  deletePageSchema,
+  fetchCommentsSchema,
+  getAboutUserSchema,
+  insertRowIntoDatabaseSchema,
+  listUsersSchema,
+  NOTION_TOOL_NAME,
+  queryDatabaseSchema,
+  retrieveBlockChildrenSchema,
+  retrieveBlockSchema,
+  retrieveDatabaseContentSchema,
+  retrieveDatabaseSchemaSchema,
+  retrievePageSchema,
+  searchSchema,
+  updatePageSchema,
+  updateRowDatabaseSchema,
+  updateSchemaDatabaseSchema,
+} from "@app/lib/actions/mcp_internal_actions/servers/notion/metadata";
 import {
   makeInternalMCPServer,
   makePersonalAuthenticationError,
@@ -31,219 +53,6 @@ import {
   parseTimeFrame,
   timeFrameFromNow,
 } from "@app/types";
-
-const uuidRegex =
-  /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/i;
-
-const parentPageSchema = z
-  .object({
-    type: z.literal("page_id").describe("Must be 'page_id' for a page parent."),
-    page_id: z.string().regex(uuidRegex).describe("The ID of the parent page."),
-  })
-  .strict();
-
-const allowedColors = [
-  "default",
-  "gray",
-  "brown",
-  "orange",
-  "yellow",
-  "green",
-  "blue",
-  "purple",
-  "pink",
-  "red",
-  "default_background",
-  "gray_background",
-  "brown_background",
-  "orange_background",
-  "yellow_background",
-  "green_background",
-  "blue_background",
-  "purple_background",
-  "pink_background",
-  "red_background",
-] as const;
-
-// We use a single tool name for monitoring given the high granularity (can be revisited).
-const NOTION_TOOL_NAME = "notion";
-
-const titleRichTextSchema = z
-  .object({
-    type: z.literal("text"),
-    text: z.object({
-      content: z.string(),
-      link: z.object({ url: z.string() }).nullable().optional(),
-    }),
-    annotations: z
-      .object({
-        bold: z.boolean().optional(),
-        italic: z.boolean().optional(),
-        strikethrough: z.boolean().optional(),
-        underline: z.boolean().optional(),
-        code: z.boolean().optional(),
-        color: z.enum(allowedColors).optional(),
-      })
-      .optional(),
-    plain_text: z.string().optional(),
-    href: z.string().nullable().optional(),
-  })
-  .strict();
-
-const propertiesSchema = z
-  .record(z.string(), z.any())
-  .default({})
-  .describe(
-    "Properties for the new page or database. Keys are property names, values are property value objects as per Notion API. See https://developers.notion.com/reference/page#property-value-types for details."
-  );
-
-const dbFilterSchema = z
-  .object({})
-  .passthrough()
-  .describe(
-    "Filter object as per Notion API. Must match the Notion API filter structure. See https://developers.notion.com/reference/post-database-query-filter"
-  );
-const dbSortSchema = z
-  .object({
-    property: z.string(),
-    direction: z.enum(["ascending", "descending"]),
-  })
-  .strict()
-  .describe(
-    "Sort object as per Notion API. See https://developers.notion.com/reference/post-database-query-sort"
-  );
-const dbSortsArraySchema = z
-  .array(dbSortSchema)
-  .describe("Array of sort objects as per Notion API.");
-
-const RichTextSchema = z
-  .object({
-    type: z.string(),
-    text: z
-      .object({
-        content: z.string(),
-        link: z.object({ url: z.string() }).nullable().optional(),
-      })
-      .optional(),
-    plain_text: z.string().optional(),
-    href: z.string().nullable().optional(),
-  })
-  .passthrough();
-
-const ParagraphBlock = z.object({
-  type: z.literal("paragraph"),
-  paragraph: z.object({
-    rich_text: z.array(RichTextSchema),
-    color: z.string().optional(),
-  }),
-  children: z.lazy(() => NotionBlockSchema.array()).optional(),
-});
-
-const Heading1Block = z.object({
-  type: z.literal("heading_1"),
-  heading_1: z.object({
-    rich_text: z.array(RichTextSchema),
-    color: z.string().optional(),
-    is_toggleable: z.boolean().optional(),
-  }),
-  children: z.lazy(() => NotionBlockSchema.array()).optional(),
-});
-
-const Heading2Block = z.object({
-  type: z.literal("heading_2"),
-  heading_2: z.object({
-    rich_text: z.array(RichTextSchema),
-    color: z.string().optional(),
-    is_toggleable: z.boolean().optional(),
-  }),
-  children: z.lazy(() => NotionBlockSchema.array()).optional(),
-});
-
-const Heading3Block = z.object({
-  type: z.literal("heading_3"),
-  heading_3: z.object({
-    rich_text: z.array(RichTextSchema),
-    color: z.string().optional(),
-    is_toggleable: z.boolean().optional(),
-  }),
-  children: z.lazy(() => NotionBlockSchema.array()).optional(),
-});
-
-const BulletedListItemBlock = z.object({
-  type: z.literal("bulleted_list_item"),
-  bulleted_list_item: z.object({
-    rich_text: z.array(RichTextSchema),
-    color: z.string().optional(),
-  }),
-  children: z.lazy(() => NotionBlockSchema.array()).optional(),
-});
-
-const NumberedListItemBlock = z.object({
-  type: z.literal("numbered_list_item"),
-  numbered_list_item: z.object({
-    rich_text: z.array(RichTextSchema),
-    color: z.string().optional(),
-  }),
-  children: z.lazy(() => NotionBlockSchema.array()).optional(),
-});
-
-const ToDoBlock = z.object({
-  type: z.literal("to_do"),
-  to_do: z.object({
-    rich_text: z.array(RichTextSchema),
-    checked: z.boolean().optional(),
-    color: z.string().optional(),
-  }),
-  children: z.lazy(() => NotionBlockSchema.array()).optional(),
-});
-
-const ToggleBlock = z.object({
-  type: z.literal("toggle"),
-  toggle: z.object({
-    rich_text: z.array(RichTextSchema),
-    color: z.string().optional(),
-  }),
-  children: z.lazy(() => NotionBlockSchema.array()).optional(),
-});
-
-const QuoteBlock = z.object({
-  type: z.literal("quote"),
-  quote: z.object({
-    rich_text: z.array(RichTextSchema),
-    color: z.string().optional(),
-  }),
-  children: z.lazy(() => NotionBlockSchema.array()).optional(),
-});
-
-const CalloutBlock = z.object({
-  type: z.literal("callout"),
-  callout: z.object({
-    rich_text: z.array(RichTextSchema),
-    icon: z.any().optional(),
-    color: z.string().optional(),
-  }),
-  children: z.lazy(() => NotionBlockSchema.array()).optional(),
-});
-
-const FallbackBlock = z
-  .object({
-    type: z.string(),
-  })
-  .passthrough();
-
-export const NotionBlockSchema: z.ZodType = z.union([
-  ParagraphBlock,
-  Heading1Block,
-  Heading2Block,
-  Heading3Block,
-  BulletedListItemBlock,
-  NumberedListItemBlock,
-  ToDoBlock,
-  ToggleBlock,
-  QuoteBlock,
-  CalloutBlock,
-  FallbackBlock,
-]);
 
 async function withNotionClient<T>(
   fn: (notion: Client) => Promise<T>,
@@ -291,21 +100,7 @@ function createServer(
   server.tool(
     "search",
     "Search for pages or databases in Notion.",
-    {
-      query: z.string().describe("Search query string."),
-      relativeTimeFrame: z
-        .string()
-        .regex(/^(all|\d+[hdwmy])$/)
-        .describe(
-          "The time frame (relative to LOCAL_TIME) to restrict the search based" +
-            " on the user request and past conversation context." +
-            " Possible values are: `all`, `{k}h`, `{k}d`, `{k}w`, `{k}m`, `{k}y`" +
-            " where {k} is a number. Be strict, do not invent invalid values."
-        ),
-      type: z
-        .enum(["page", "database"])
-        .describe("What type of notion objects to search."),
-    },
+    searchSchema,
     withToolLogging(
       auth,
       {
@@ -454,9 +249,7 @@ function createServer(
   server.tool(
     "retrieve_page",
     "Retrieve a Notion page by its ID.",
-    {
-      pageId: z.string().describe("The Notion page ID."),
-    },
+    retrievePageSchema,
     withToolLogging(
       auth,
       {
@@ -475,9 +268,7 @@ function createServer(
   server.tool(
     "retrieve_database_schema",
     "Retrieve a Notion database's schema by its ID.",
-    {
-      databaseId: z.string().describe("The Notion database ID."),
-    },
+    retrieveDatabaseSchemaSchema,
     withToolLogging(
       auth,
       {
@@ -496,16 +287,7 @@ function createServer(
   server.tool(
     "retrieve_database_content",
     "Retrieve the content (pages) of a Notion database by its ID.",
-    {
-      databaseId: z.string().describe("The Notion database ID."),
-      filter: dbFilterSchema.optional(),
-      sorts: dbSortsArraySchema.optional(),
-      start_cursor: z
-        .string()
-        .optional()
-        .describe("Start cursor for pagination."),
-      page_size: z.number().optional().describe("Page size for pagination."),
-    },
+    retrieveDatabaseContentSchema,
     withToolLogging(
       auth,
       {
@@ -534,16 +316,7 @@ function createServer(
   server.tool(
     "query_database",
     "Query a Notion database.",
-    {
-      databaseId: z.string().describe("The Notion database ID."),
-      filter: dbFilterSchema.optional(),
-      sorts: dbSortsArraySchema.optional(),
-      start_cursor: z
-        .string()
-        .optional()
-        .describe("Start cursor for pagination."),
-      page_size: z.number().optional().describe("Page size for pagination."),
-    },
+    queryDatabaseSchema,
     withToolLogging(
       auth,
       {
@@ -572,14 +345,7 @@ function createServer(
   server.tool(
     "create_page",
     "Create a new Notion page.",
-    {
-      parent: parentPageSchema.describe(
-        "The existing parent page where the new page is inserted. Must be an existing page ID. Use the search action to find a page ID."
-      ),
-      properties: propertiesSchema,
-      icon: z.any().optional().describe("Icon (optional)."),
-      cover: z.any().optional().describe("Cover (optional)."),
-    },
+    createPageSchema,
     withToolLogging(
       auth,
       {
@@ -598,12 +364,7 @@ function createServer(
   server.tool(
     "insert_row_into_database",
     "Create a new Notion page in a database.",
-    {
-      databaseId: z.string().describe("The Notion database ID."),
-      properties: propertiesSchema,
-      icon: z.any().optional().describe("Icon (optional)."),
-      cover: z.any().optional().describe("Cover (optional)."),
-    },
+    insertRowIntoDatabaseSchema,
     withToolLogging(
       auth,
       {
@@ -628,19 +389,7 @@ function createServer(
   server.tool(
     "create_database",
     "Create a new Notion database (table).",
-    {
-      parent: parentPageSchema.describe(
-        "The existing parent where the new database is inserted. Must be an existing page ID. Use the search action to find a page ID."
-      ),
-      title: z
-        .array(titleRichTextSchema)
-        .describe(
-          "Title for the database as an array of rich text objects (see Notion API). Each item must have type: 'text' and a text object with content."
-        ),
-      properties: propertiesSchema,
-      icon: z.any().optional().describe("Icon (optional)."),
-      cover: z.any().optional().describe("Cover (optional)."),
-    },
+    createDatabaseSchema,
     withToolLogging(
       auth,
       {
@@ -660,10 +409,7 @@ function createServer(
   server.tool(
     "update_page",
     "Update a Notion page's properties.",
-    {
-      pageId: z.string().describe("The Notion page ID."),
-      properties: propertiesSchema,
-    },
+    updatePageSchema,
     withToolLogging(
       auth,
       {
@@ -682,9 +428,7 @@ function createServer(
   server.tool(
     "retrieve_block",
     "Retrieve a Notion block by its ID.",
-    {
-      blockId: z.string().describe("The Notion block ID."),
-    },
+    retrieveBlockSchema,
     withToolLogging(
       auth,
       {
@@ -703,14 +447,7 @@ function createServer(
   server.tool(
     "retrieve_block_children",
     "Retrieve the children of a Notion block or page by its ID.",
-    {
-      blockId: z.string().describe("The Notion block or page ID."),
-      start_cursor: z
-        .string()
-        .optional()
-        .describe("Start cursor for pagination."),
-      page_size: z.number().optional().describe("Page size for pagination."),
-    },
+    retrieveBlockChildrenSchema,
     withToolLogging(
       auth,
       {
@@ -734,14 +471,7 @@ function createServer(
   server.tool(
     "add_page_content",
     "Add a single content block to a Notion page. For multiple blocks, call this action multiple times. Only supports adding to Notion pages. Blocks that can contain children include: page, toggle, to-do, bulleted list, numbered list, callout, and quote.",
-    {
-      blockId: z.string().describe("The ID of the parent block or page."),
-      children: z
-        .array(NotionBlockSchema)
-        .describe(
-          "Array of block objects to append as children. Blocks can be parented by other blocks, pages, or databases. There is a limit of 100 block children that can be appended by a single API request."
-        ),
-    },
+    addPageContentSchema,
     withToolLogging(
       auth,
       {
@@ -764,22 +494,7 @@ function createServer(
   server.tool(
     "create_comment",
     "Create a comment on a Notion page or in an existing discussion thread. Provide either a parent page ID or a discussion ID. Inline comments to start a new thread are not supported via the public API.",
-    {
-      parent_page_id: z
-        .string()
-        .regex(uuidRegex)
-        .optional()
-        .describe(
-          "The ID of the parent page (optional, required if not using discussion_id)."
-        ),
-      discussion_id: z
-        .string()
-        .optional()
-        .describe(
-          "The ID of the discussion thread (optional, required if not using parent_page_id)."
-        ),
-      comment: z.string().describe("The comment text."),
-    },
+    createCommentSchema,
     withToolLogging(
       auth,
       {
@@ -814,9 +529,7 @@ function createServer(
   server.tool(
     "delete_block",
     "Archive (delete) block content in a page. In the Notion UI, this moves the block to the 'trash,' where it can be restored if needed.",
-    {
-      blockId: z.string().describe("The ID of the block"),
-    },
+    deleteBlockSchema,
     withToolLogging(
       auth,
       {
@@ -836,9 +549,7 @@ function createServer(
   server.tool(
     "delete_page",
     "Archive (delete) a page or database row. In the Notion UI, this moves the block to the 'trash,' where it can be restored if needed.",
-    {
-      pageId: z.string().describe("The ID of the page"),
-    },
+    deletePageSchema,
     withToolLogging(
       auth,
       {
@@ -857,11 +568,7 @@ function createServer(
   server.tool(
     "fetch_comments",
     "Retrieve a list of unresolved comment objects from a specified page or block in Notion.",
-    {
-      blockId: z
-        .string()
-        .describe("The ID of the page or block to fetch comments from."),
-    },
+    fetchCommentsSchema,
     withToolLogging(
       auth,
       {
@@ -880,10 +587,7 @@ function createServer(
   server.tool(
     "update_row_database",
     "Update a specific property value in a row (page) of a Notion database. Value formats depend on the property type (e.g., text, number, select, date, people, URL, files, checkbox).",
-    {
-      pageId: z.string().regex(uuidRegex).describe("The Notion page ID."),
-      properties: propertiesSchema,
-    },
+    updateRowDatabaseSchema,
     withToolLogging(
       auth,
       {
@@ -902,10 +606,7 @@ function createServer(
   server.tool(
     "update_schema_database",
     "Update the schema (columns/properties) of an existing Notion database.",
-    {
-      databaseId: z.string().describe("The Notion database ID."),
-      properties: propertiesSchema,
-    },
+    updateSchemaDatabaseSchema,
     withToolLogging(
       auth,
       {
@@ -925,7 +626,7 @@ function createServer(
   server.tool(
     "list_users",
     "List all users in the Notion workspace.",
-    {},
+    listUsersSchema,
     withToolLogging(
       auth,
       {
@@ -941,9 +642,7 @@ function createServer(
   server.tool(
     "get_about_user",
     "Get information about a specific user by userId.",
-    {
-      userId: z.string().describe("The Notion user ID."),
-    },
+    getAboutUserSchema,
     withToolLogging(
       auth,
       {
